refactored hooks into two
