+++ Client/src/hooks/index.ts
@@ -3,4 +3,5 @@ export * from './use-graph-data'
 export * from './use-mobile'
 export * from './use-outside-click'
 export * from './use-subscribe-to-topic'
-export * from './use-device-data'
+export * from './use-device'
+export * from './use-device-sensor-data'
